@@ -1,29 +1,13 @@
 {
-	"name": "@radiofrance/svelte-leaflet",
-	"version": "0.1.0-alpha.19",
-	"description": "A library that wraps leaflet classes in domless/renderless svelte components.",
-	"keywords": [
-		"svelte",
-		"leaflet",
-		"map",
-		"cartography",
-		"sveltekit"
-	],
-	"author": {
-		"email": "romain.durand@radiofrance.com",
-		"name": "Romain Durand"
-	},
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/radiofrance/svelte-leaflet"
-	},
+	"name": "svelte-leaflet2",
+	"version": "0.0.1",
 	"scripts": {
 		"dev": "vite dev",
-		"build": "vite build && npm package",
+		"build": "vite build && npm run package",
 		"preview": "vite preview",
 		"package": "svelte-kit sync && svelte-package && publint",
 		"prepublishOnly": "npm run package",
-		"test": "npm run check && npm run test:unit",
+		"test": "npm run test:integration && npm run test:unit",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
 		"lint": "prettier --check . && eslint .",
@@ -43,59 +27,34 @@
 		"!dist/**/*.spec.*"
 	],
 	"peerDependencies": {
-		"svelte": "^4.0.0"
+		"svelte": "^5.0.0-next.1"
 	},
 	"devDependencies": {
-		"@csstools/postcss-global-data": "^2.1.1",
-		"@fontsource-variable/fira-code": "^5.0.18",
 		"@playwright/test": "^1.28.1",
-		"@shikijs/core": "1.12.1",
-		"@shikijs/transformers": "^1.10.3",
-		"@shikijs/twoslash": "^1.12.1",
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "^2.0.0",
 		"@sveltejs/package": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/d3-scale": "^4.0.8",
-		"@types/d3-scale-chromatic": "^3.0.3",
-		"@types/eslint": "8.56.0",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"autoprefixer": "^10.4.19",
-		"d3-scale": "^4.0.2",
-		"d3-scale-chromatic": "^3.0.0",
-		"eslint": "^8.56.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
+		"@types/eslint": "^9.6.0",
+		"@types/leaflet": "^1.9.12",
+		"@types/leaflet.markercluster": "^1.5.4",
+		"eslint": "^9.0.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"mdsvex": "^0.11.2",
-		"node-html-parser": "^6.1.13",
-		"postcss": "^8.4.39",
-		"postcss-csso": "^6.0.1",
-		"postcss-preset-env": "^9.6.0",
+		"eslint-plugin-svelte": "^2.36.0",
+		"globals": "^15.0.0",
+		"leaflet": "^1.9.4",
+		"leaflet.markercluster": "^1.5.3",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
-		"publint": "^0.1.9",
-		"shiki": "^1.10.3",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"svelte-preprocess": "^6.0.1",
-		"tslib": "^2.4.1",
+		"publint": "^0.2.0",
+		"svelte": "^5.0.0-next.1",
+		"svelte-check": "^4.0.0",
 		"typescript": "^5.0.0",
+		"typescript-eslint": "^8.0.0",
 		"vite": "^5.0.11",
-		"vitest": "^1.6.0"
+		"vitest": "^2.0.0"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
-	"type": "module",
-	"private": false,
-	"publishConfig": {
-		"access": "public",
-		"registry": "https://registry.npmjs.org/"
-	},
-	"dependencies": {
-		"@types/leaflet": "^1.9.12",
-		"@types/leaflet.markercluster": "^1.5.4",
-		"leaflet": "^1.9.4",
-		"leaflet.markercluster": "^1.5.3"
-	}
+	"type": "module"
 }
