@@ -19,14 +19,15 @@ import type { EventDispatcher } from 'svelte';
 
 // Reexport your entry components here
 export { default as Map } from './Map.svelte';
-export { default as Marker } from './Marker.svelte';
-export { default as MarkerClusterGroup } from './MarkerClusterGroup.svelte';
-export { default as Polyline } from './Polyline.svelte';
-export { default as Popup } from './Popup.svelte';
-export { default as Circle } from './Circle.svelte';
-export { default as Polygon } from './Polygon.svelte';
+// export { default as Marker } from './Marker.svelte';
+// export { default as MarkerClusterGroup } from './MarkerClusterGroup.svelte';
+// export { default as Polyline } from './Polyline.svelte';
+// export { default as Popup } from './Popup.svelte';
+// export { default as Circle } from './Circle.svelte';
+// export { default as Polygon } from './Polygon.svelte';
 
-export type { Marker as LeafletMarker } from './Marker.svelte';
+// export type { Marker as LeafletMarker } from './Marker.svelte';
+
 export type {
 	Circle as LeafletCircle,
 	CircleOptions,
